refactor(codecs): inherit iter_xpaths from CodecBase, drop 9 sibling-import overrides
Fable review 2026-07-06 (finding #64).

The canonical-tree xpath walker was relocated to canonical/xpath_walker.py to
escape the vendor packages, yet 11 codecs still reached it THROUGH the
cisco_iosxe_cli package (`from ..cisco_iosxe_cli.codec import _walk_canonical`)
via a copy-pasted iter_xpaths override — the exact cross-vendor edge the
relocation meant to remove, and the pattern the README taught codec #13.

- CodecBase.iter_xpaths now walks a CanonicalIntent via the shared
  canonical.xpath_walker (runtime-imported; keeps the flat mock-dict branch).
- Deleted the 9 byte-identical overrides (arista_eos / aruba_aoscx /
  aruba_aoss / cisco_iosxr / cisco_nxos / fortigate_cli / juniper_junos /
  mikrotik_routeros / vyos) — they now inherit; removed the orphaned Iterable
  imports.
- cisco_iosxe (XML) and opnsense KEEP their overrides (genuine nested-dict
  paths) but import _walk_canonical from canonical.xpath_walker, not the
  sibling package.
- cisco_iosxe_cli keeps its module-level re-export for back-compat.
- README (3 spots) + 4 codec module-docstrings corrected — they taught/claimed
  the legacy sibling-import path.

Behaviour-preserving: every codec's iter_xpaths yields the same xpaths (same
walker, called from base) — sanity-checked across all 13 codecs. Full
tests/unit + cross-mesh guard green (5200 passed); the walker-completeness +
capability-honesty guards are the proof. ruff clean.

Note: the finding said "11 identical overrides"; two (cisco_iosxe XML,
opnsense) have genuine nested-dict paths and are kept (import swapped), so 9
were deleted.
