+++ netcanon/migration/codecs/cisco_nxos/codec.py
@@ -17,10 +17,9 @@ Module layout mirrors the ``cisco_iosxe_cli`` post-split shape:
 * ``render.py`` — canonical tree → NX-OS running-config text.
 * ``port_names.py`` — cross-vendor port-name bridge.
 
-``iter_xpaths`` reuses ``_walk_canonical`` from
-:mod:`netcanon.migration.codecs.cisco_iosxe_cli.codec` — NX-OS
-introduces no new canonical xpaths in Phase 1, so the shared walker
-yields exactly the right set.
+``iter_xpaths`` is inherited from :class:`CodecBase`, which walks a
+``CanonicalIntent`` via the shared ``canonical.xpath_walker`` — NX-OS
+introduces no new canonical xpaths, so no override is needed (#64).
 
 This codec lands in four phases (see
 ``docs/v0.2.0-planning/03-nxos-codec/``).  **All four phases are
@@ -47,7 +46,6 @@ L2VNI, EVPN L3VNI, BGP-redistribute) in
 from __future__ import annotations
 
 import re
-from collections.abc import Iterable
 from typing import Any, ClassVar
 
 from ....models.migration import (
@@ -607,16 +605,6 @@ class CiscoNXOSCodec(CodecBase):
 
     def render(self, tree: Any) -> str:
         return render_intent(tree)
-
-    # -----------------------------------------------------------------
-    # iter_xpaths — reuse the shared canonical walker
-    # -----------------------------------------------------------------
-
-    def iter_xpaths(self, tree: Any) -> Iterable[str]:
-        """Yield schema xpaths from a :class:`CanonicalIntent`."""
-        if isinstance(tree, CanonicalIntent):
-            from ..cisco_iosxe_cli.codec import _walk_canonical
-            yield from _walk_canonical(tree)
 
     # -----------------------------------------------------------------
     # Cross-vendor port-name translation (delegated to .port_names)
